Fix auto-complete malfunction after enable `web`
The idea is to maintain a copy of the `struct line_state l` in list_edit
as a static global structure. When the socket is available for reading,
I copy the list_status from l to g. After handling the command from the
socket, restore the status back to l from g.